Update readability.c
Modify readability.c so that, instead of printing out the literal text itself, it instead prints out a count of the number of letters in the text.

Uses a function called count_letters. This function uses isalpha to determine whether a character is alphabetical. 
If character is alphabetical, 1 is added to letter count. 
If not alphabetical, 0 is added to letter count.

'count' is then returned and assigned to 'letters', printing out the number of letters in the inputted text.

The function returns an integer representing the number of letters within the inputted text and this is printed